update collections and landing pages templates
If the user clicks on "Preview" button or "View Draft" button, they
expect to see the *latest* version of a page.

If the user clicks on "Status" button or "View Live" button, they expect
to see the latest *published* version of a page.

In this commit, these expectations are fulfilled by checking the URL
pattern of Wagtail and redirecting to frontend of tapahtumat helsinki
with the draft flag set correctly depending on which button the user has
clicked on.