overwrite first Finnish translation for a Wagtail message
The Wagtail message of:

"The page could not be saved due to validation errors"

Is not understandable enough for our content creators. This had to be
overwritten to:

"Syöttämisässi tiedoissa on virhe"

For this, application/wagtail_messages.py has been added with
instructions for future overwriting of messages as well.