implement logic for putting pages on top after their creation
when content creators create a page, let's say a Collection page,
by default wagtail adds that page to the bottom of the list of
Collection pages. this is not the desired behaviour for content creators
of tapahtumat.helsinki and therefore we needed to change it.

after much investigation, I found that this is elegantly possible by
using the after_create_page hook which comes by Wagtail and just moving
the page to the desired location using the move() function which is
available on every page instance.