enable content editors to modify slug for collections
slug was previously autogenerated not to confuse content creators with
such technical jargon. we now have the need however to show real words
in collections url for SEO, and therefore content editors must be able
to have control over slug. so I removed the autogeration logic for slug
and added a field panel so they can edit it.